requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
PyPDF2
python-dotenv
jinja2